Bit of a downgrade
The new open solution did not work, so I went with the tried and true
defining encoding everytime.

Which has led to some errors in Python 3, at the moment I've decided to
work exclusively in Python 2.7. When I have core code I'm happy with,
I'll work about compatability.

Also, as per a conversation in the comments over in Reddit, I built
branch called Scrypto (will be posted on GitHub after this.) You can
read about Scrypto in the change log.